requirements.txt
streamlit
spotipy
requests
python-dotenv
pandas
plotly
pyinstaller